Add files via upload
Using chi-squared analysis, the optimized gradients for each height of water were found using this code, for the second tube. Please not, the code for calculating chi-squared and the reduced chi-squared, as well as the optimized parameters was provided, and was not my own code, although I did adjust it slightly in places.