Add files via upload
This is the chi-squared analysis for the second tube. Please not, the code for calculating chi-squared and the reduced chi-squared, as well as the optimized parameters was provided, and was not my own code, although I did adjust it slightly in places. Additionally, for the code for the contour plot, I used generative AI for some help with coding this.